requirements.txt
autopep8==1.5.7
Cerberus==1.3.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
pycryptodome==3.10.1
PyJWT==2.1.0
PyMySQL==1.0.2
python-dotenv==0.17.1
toml==0.10.2
Werkzeug==2.0.1
